backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
greenlet>=3.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation / Settings
pydantic>=2.7.0
pydantic-settings>=2.3.0

# AI / OCR
anthropic>=0.28.0

# File Processing
openpyxl>=3.1.0
pandas>=2.2.0
python-docx>=1.1.0
pdf2image>=1.17.0
PyMuPDF>=1.24.0
Pillow>=10.0.0

# Storage
boto3>=1.34.0

# Utilities
python-dateutil>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
xlrd>=2.0.0
email-validator>=2.1.0

# Dev / Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.5.0
isort>=5.13.0
